Fix warnings, apply patch only for linux kernels>=4.10
The nvidia-drivers ebuild takes into account freeBSD
kernels as well. To prevent unforseen consequences the
patch is only applied if the linux kernel is >=4.10.

Also, the warning messages in pkg_pretend() were updated
to reflect the patched kernel which should be able to work.

Ebuild hasn't been tested on a live system.