Fix blunder of not including patchfile
I managed to forget to include a patch
by pushing a commit from a different
location than I used for testing the
tigervnc-1.7.1-r2 ebuild. So the ebuild
worked in testing, but the patch wasn't
included.

Also, I fixed the naming scheme for the
inetd patch to be more consistent.