sys-apps/portage Fl-3687 add git branch control
Define a sync-branch variable in repos.conf that
can be used to clone and checkout a certain git
branch. It can be used with other CVS-es, but
a proof-of-concept implementation is only given
here for git.

I haven't found an elegant way to clone all branches
and sync them: this will require a lot more work
than syncing only one branch. I'd prefer that this
kind of functionality be provided by an external tool,
and not portage: an ego/eselect module could do the
trick here.

git wasn't made for package management, and this becomes
obvious when trying to implement some simple features.

Simply put, this patch (portage-add-sync-branch.patch)
allows one to set an initial branch that should
be used with portage (e.g. for generating a stage3 or stage4,
or for setting a safe default for first time users).

The only thing I think portage should do while doing the initial
cloning is to fetch all the remote branches (just the branch
information, not the actual branch data). This should allow other,
more advanced overlay/repository management tools to more easily
fetch/switch between branches.