Add ebuilds for dia and mpaste
mpaste is a python pastebin client that interfaces
to my Funtoo hosted pastebin service paste.mheaven.ch
It's based on the pasteme pastit.py file.

dia is a basic diagnostic script that gathers system
information and posts it using mpaste to paste.mheaven.ch.
It's designed to simplify Funtoo support.