FL-3785: fixed by upgrading standard to std=c++11
The issue is present in all gcc compilers lower than
version 6, since features that need the c++11 standard
of c++ are required to compile this ebuild.

The fix is to append the -std=c++11 flag to cxxflags
to force adherence to the newer standard. The default
standard for gcc lower than version 6 is c++98.