Working skeleton ebuild for mail-client/s-nail
The ebuild pulls in the specified dependencies that
are required for a standard build.

The ebuild inherits a lot of details from nail, which
is the origin of the fork, and therefore, both ebuilds
share similarities on how the configure and compile phases
are handled.

The ebuild seems to install correctly.

Implementing proper use flags (ssl, pop3, imap) can be done in
a later version. Everything except GSS-API authentication and
Terminal capability queries is turned on in the default install,
so the need to add use flags is small. Feel free to hack together
a better ebuild if you have the time.