FL-3871: Restore sys-firmware/iwl4965-ucode
THe ebuilds have been probably treecleaned
due to the fact that the SRC_URI changed,
and the firmwares were unavailable for
a while before being moved to their
current location. Therefore, Gentoo
tree acleaned the ebuild and recommended
users to switch to  linux-firmware..

I updated the SRC_URI and fixed some
inconsistencies in the ebuilds, and
updated the copyright info.

The sources are also mirrored for good
measure on the distmirror.ceresia.ch
distfiles.ceresia.ch . The Intel firmware
license allows mirroring :)

The firmware is also available through
linux-firmware. The firmware in linux-firmware
is identical to the latest ebuild added here
(228.61.2.24) which I'm running with the
latest gentoo-sources-4.11.6 kernel. Works
as expected.